Implemented h_setlevel heuristic function
Implemented the h_setlevel heuristic function in an
inefficient manner. The function fills the graph first
and searches for the state in which all goals are
satisfied, and none of the goals mutually exclusive.

Further refinement can be done to improve efficiency by
expanding the graph at each check instead of filling it,
similar to the reccommendations for the h_levelsum
heuristic function.